refactor(llm): group agent review internals
